sql-based post-processing

+++ src/main/java/com/feiliks/dashboard/spring/ConsolidationController.java
@@ -32,7 +32,7 @@ public class ConsolidationController {
     }
 
     @GetMapping("/status.json")
-    public ResponseEntity<Map<String, Integer>> getStatus() {
+    public ResponseEntity<Map<String, Long>> getStatus() {
         return ResponseEntity.ok(task.getCurrentStats());
     }
